+++ commands/music.js
@@ -4,11 +4,11 @@ module.exports = {
     data: new SlashCommandBuilder()
         .setName('music')
         .setDescription('Open music GUI'),
-    execute(client, dmobj, message) {
+    execute(client, dmobj, args) {
 
         dmobj.set_client(client);
-        dmobj.set_last_at_channel(message.channel);
-        dmobj.join_channel(message);
-        dmobj.send_control_panel(message);
+        dmobj.set_last_at_channel(args.channel);
+        dmobj.join_channel(args);
+        dmobj.send_control_panel(args);
     },
 };