+++ commands/ping.js
@@ -4,7 +4,7 @@ module.exports = {
     data: new SlashCommandBuilder()
         .setName('ping')
         .setDescription('Replies with Pong!'),
-    async execute(client, message, args) {
-        return message.channel.send('Pong!');
+    async execute(client, args, argsStr) {
+        return args.channel.send('Pong!');
     },
 };