+++ events/messageCreate.js
@@ -1,22 +1,10 @@
-const fs = require('fs');
-const { Collection } = require('discord.js');
 const { globalPrefix } = require('../config.json');
 //var memwatch = require('memwatch-next');
 //var heapdump = require('heapdump');
-const commands = new Collection();
-
-const commandFiles = fs.readdirSync('./commands').filter(file => file.endsWith('.js'));
-
-for (const file of commandFiles) {
-    const command = require(`../commands/${file}`);
-    // Set a new item in the Collection
-    // With the key as the command name and the value as the exported module
-    commands.set(command.data.name, command);
-}
 
 module.exports = {
     name: 'messageCreate',
-    async execute(client, dmobj, message) {
+    async execute(client, dmobj, commands, message) {
         //console.log(` sent a message.`);
 
         let is_command = false;
@@ -49,17 +37,9 @@ module.exports = {
 
 
         if (!is_command) return;
-
-        const command = commands.get(command_str);
         try {
-            if (command_str === "music") {
-                command.execute(client, dmobj, message, args);
-            } else if (command_str === "memsnap") {
-                //heapdump.writeSnapshot();
-            } else {
-                command.execute(client, message, args);
-            }
 
+            commands.executeDiscordCommand(command_str, message, args);
         } catch (error) {
             console.error(error);
             message.channel.send({ content: 'There was an error while executing this command!', ephemeral: true });